% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_class.R
\name{make_class}
\alias{make_class}
\title{make plant UML for a class}
\usage{
make_class(r6class)
}
\arguments{
\item{r6class}{an R6 class generator}
}
\value{
text of the UML
}
\description{
Create UML for a class
}
\author{
Barry Rowlingson
}
