% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_class.R
\name{make_classes}
\alias{make_classes}
\title{Plant UML for a list of classes}
\usage{
make_classes(classlist)
}
\arguments{
\item{classlist}{a list of R6 class generators}
}
\value{
UML text
}
\description{
Create UML for a list of classes
}
\author{
Barry Rowlingson
}
