% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plant.R
\name{make_plant}
\alias{make_plant}
\title{Create UML diagram from R6 object generators}
\usage{
make_plant(classes, output = stdout())
}
\arguments{
\item{classes}{list of R6 class generators}

\item{output}{filename or connection for output}
}
\value{
none
}
\description{
Make PlantUML diagram from a list of R6 classes
}
\author{
Barry Rowlingson
}
